backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0

# LangChain + RAG
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# PDF processing
pypdf>=4.0.0

# Vector DB
chromadb>=0.5.0

# Integrations
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
supabase>=2.0.0
python-telegram-bot>=20.0
